+++ oshi-core/src/main/java/oshi/util/platform/windows/PerfCounterQuery.java
@@ -24,10 +24,9 @@
 package oshi.util.platform.windows;
 
 import java.util.EnumMap;
-import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.locks.ReentrantLock;
+import java.util.concurrent.ConcurrentHashMap;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,7 +39,6 @@ import com.sun.jna.platform.win32.COM.Wbemcli;
 import com.sun.jna.platform.win32.COM.WbemcliUtil.WmiQuery;
 import com.sun.jna.platform.win32.COM.WbemcliUtil.WmiResult;
 
-import oshi.annotation.concurrent.GuardedBy;
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.util.platform.windows.PerfDataUtil.PerfCounter;
 
@@ -54,10 +52,12 @@ public final class PerfCounterQuery {
 
     private static final boolean IS_VISTA_OR_GREATER = VersionHelpers.IsWindowsVistaOrGreater();
 
-    // Use a map to cache failed pdh queries
-    @GuardedBy("failedQueryCacheLock")
-    private static final Set<String> failedQueryCache = new HashSet<>();
-    private static final ReentrantLock failedQueryCacheLock = new ReentrantLock();
+    // Use a thread safe set to cache failed pdh queries
+    private static final Set<String> failedQueryCache = ConcurrentHashMap.newKeySet();
+
+    // For XP, use a map to cache localization strings
+    private static final ConcurrentHashMap<String, String> localizeCache = IS_VISTA_OR_GREATER ? null
+            : new ConcurrentHashMap<>();
 
     /*
      * Multiple classes use these constants
@@ -91,23 +91,14 @@ public final class PerfCounterQuery {
      */
     public static <T extends Enum<T>> Map<T, Long> queryValues(Class<T> propertyEnum, String perfObject,
             String perfWmiClass) {
-        // Check without locking for performance
         if (!failedQueryCache.contains(perfObject)) {
-            failedQueryCacheLock.lock();
-            try {
-                // Double check lock
-                if (!failedQueryCache.contains(perfObject)) {
-                    Map<T, Long> valueMap = queryValuesFromPDH(propertyEnum, perfObject);
-                    if (!valueMap.isEmpty()) {
-                        return valueMap;
-                    }
-                    // If we are here, query failed
-                    LOG.warn("Disabling further attempts to query {}.", perfObject);
-                    failedQueryCache.add(perfObject);
-                }
-            } finally {
-                failedQueryCacheLock.unlock();
+            Map<T, Long> valueMap = queryValuesFromPDH(propertyEnum, perfObject);
+            if (!valueMap.isEmpty()) {
+                return valueMap;
             }
+            // If we are here, query failed
+            LOG.warn("Disabling further attempts to query {}.", perfObject);
+            failedQueryCache.add(perfObject);
         }
         return queryValuesFromWMI(propertyEnum, perfWmiClass);
     }
@@ -132,16 +123,14 @@ public final class PerfCounterQuery {
     public static <T extends Enum<T>> Map<T, Long> queryValuesFromPDH(Class<T> propertyEnum, String perfObject) {
         T[] props = propertyEnum.getEnumConstants();
         // If pre-Vista, localize the perfObject
-        if (!IS_VISTA_OR_GREATER) {
-            perfObject = PerfCounterQuery.localize(perfObject);
-        }
+        String perfObjectLocalized = PerfCounterQuery.localizeIfNeeded(perfObject);
         EnumMap<T, PerfCounter> counterMap = new EnumMap<>(propertyEnum);
         EnumMap<T, Long> valueMap = new EnumMap<>(propertyEnum);
         try (PerfCounterQueryHandler pdhQueryHandler = new PerfCounterQueryHandler()) {
             // Set up the query and counter handles
             for (T prop : props) {
-                PerfCounter counter = PerfDataUtil.createCounter(perfObject, ((PdhCounterProperty) prop).getInstance(),
-                        ((PdhCounterProperty) prop).getCounter());
+                PerfCounter counter = PerfDataUtil.createCounter(perfObjectLocalized,
+                        ((PdhCounterProperty) prop).getInstance(), ((PdhCounterProperty) prop).getCounter());
                 counterMap.put(prop, counter);
                 if (!pdhQueryHandler.addCounterToQuery(counter)) {
                     return valueMap;
@@ -213,7 +202,12 @@ public final class PerfCounterQuery {
      * @return The localized string if localization successful, or the original
      *         string otherwise.
      */
-    public static String localize(String perfObject) {
+    public static String localizeIfNeeded(String perfObject) {
+        return IS_VISTA_OR_GREATER ? perfObject
+                : localizeCache.computeIfAbsent(perfObject, PerfCounterQuery::localizeUsingPerfIndex);
+    }
+
+    private static String localizeUsingPerfIndex(String perfObject) {
         String localized = perfObject;
         try {
             localized = PdhUtil.PdhLookupPerfNameByIndex(null, PdhUtil.PdhLookupPerfIndexByEnglishName(perfObject));